refactor: cleanup cli interface

+++ cli/process.js
@@ -1,50 +1,12 @@
 import fs from "fs";
-import config from "nconf";
 import path from "path";
+import yargs from "yargs";
+import { hideBin } from "yargs/helpers";
 
 import { splitTextStyleAndMetadata } from "../shared/helpers.js";
 import markdown from "../shared/naturalcrit/markdown.js";
 
-process.chdir(path.resolve());
-
-config.argv({ x: { lowerCase: true } }).file({ file: "config.json" });
-
-// Check mandatory parameters exist
-const requiredParams = ["input", "output", "renderer"];
-const params = requiredParams.map((param) => {
-	return config.get(param);
-});
-if (
-	params.some((x) => {
-		return !x;
-	})
-) {
-	console.log("Required parameter missing");
-	requiredParams.forEach((param) => {
-		console.log(`${param} : ${config.get(param)}`);
-	});
-	console.log("Please check your parameters before trying again.");
-}
-
-// Check if output file exists, and if we're allowed to overwrite it
-if (fs.existsSync(config.get("output")) && !config.get("overwrite")) {
-	console.log(
-		"Specified output file exists: please specify --overwrite to replace.",
-	);
-	process.exit();
-}
-
-// Read input file
-const brew = {
-	text: fs.readFileSync(config.get("input"), { encoding: "UTF-8" }),
-};
-
-// Parse brew text to populate brew object
-// This mutates the passed object
-splitTextStyleAndMetadata(brew);
-
-// Set Renderer Options
-const RendererOptions = {
+const RENDER_OPTIONS = {
 	// NOTE: legacy renderer is unsupported
 	// legacy: {
 	// module: markdownLegacy,
@@ -64,22 +26,30 @@ const RendererOptions = {
 	},
 };
 
-const Marked = RendererOptions[config.get("renderer")].module;
+const parseBrewText = (text) => {
+	const brew = { text };
+	splitTextStyleAndMetadata(brew);
+	return brew;
+};
 
-// Initialize a list to render the pages in to
-brew.html = [];
+const renderBrewPages = (text, renderer) =>
+	text
+		// split into markdown pages
+		.split(renderer.pageRegex)
+		// render to html pages
+		.map(
+			(page, index) =>
+				`
+        <div class='page phb' id='p${index + 1}' key='${index}' ${renderer.divHeader}>
+          ${renderer.module.render(page)}${renderer.divFooter}\n
+        </div>\n`,
+		);
 
-// Split brew.text into pages and render to HTML
-const pages = brew.text.split(
-	RendererOptions[config.get("renderer")].pageRegex,
-);
-pages.forEach((page, index) => {
-	brew.html[index] =
-		`<div class='page phb' id='p${index + 1}' key='${index}' ${RendererOptions[config.get("renderer")].divHeader}>${Marked.render(page)}${RendererOptions[config.get("renderer")].divFooter}\n</div>\n`;
-});
+const renderHTML = (brew, renderer) => {
+	const pages = renderBrewPages(brew.text, renderer);
 
-// Wrap the output in a HTML template
-const htmlOutput = `<!DOCTYPE html>
+	// Wrap the output in a HTML template
+	const htmlOutput = `<!DOCTYPE html>
 	<html>
 		<head>
 			<link href="https://use.fontawesome.com/releases/v5.15.1/css/all.css" rel="stylesheet" />
@@ -89,22 +59,78 @@ const htmlOutput = `<!DOCTYPE html>
 			<title>The Homebrewery - Local Output</title>
 		</head>
 		<body>
-      ${RendererOptions[config.get("renderer")].pageHeader}
+      ${renderer.pageHeader}
       <div class='brewRenderer'>
         <style>${brew.style}</style>
         <div class='pages'>
-			    ${brew.html.join("\n")}
+			    ${pages.join("\n")}
         </div>
       </div>
 		</body>
 	</html>
 	`;
+	return htmlOutput;
+};
 
-const outputDir = path.dirname(config.get("output"));
-fs.mkdirSync(outputDir, { recursive: true });
-// Write everything to the output file
-fs.writeFileSync(config.get("output"), htmlOutput);
-// Write supporting web assets
-fs.cpSync("./build/", path.join(outputDir, "./"), { recursive: true });
+const render = (brewText, rendererVersion) => {
+	const renderer = RENDER_OPTIONS[rendererVersion];
+	const brew = parseBrewText(brewText);
+	const html = renderHTML(brew, renderer);
 
-console.log(`Output written to file: ${config.get("output")}`);
+	return html;
+};
+
+yargs(hideBin(process.argv))
+	.scriptName("homebrewery")
+	.usage("$0 <cmd> [args]")
+	.command(
+		"write-brew <input> <output>",
+		"write brew to file headlessly",
+		(yargs) => {
+			yargs.positional("input", {
+				type: "string",
+				describe: "path to markdown brew file",
+			});
+			yargs.positional("output", {
+				type: "string",
+				describe: "path to html output file",
+			});
+			yargs.option("r", {
+				alias: "renderer",
+				default: "v3",
+				describe: "Renderer to use. Only v3 is supported.",
+				type: "string",
+			});
+			yargs.option("x", {
+				alias: "overwrite",
+				default: false,
+				describe: "Should overwrite existing files",
+				type: "boolean",
+			});
+		},
+		function (argv) {
+			// Check if output file exists, and if we're allowed to overwrite it
+			if (fs.existsSync(argv.output) && !argv.overwrite) {
+				console.log(
+					"Specified output file exists: please specify --overwrite to replace.",
+				);
+				process.exit();
+			}
+
+			// Read input file
+			const inputText = fs.readFileSync(argv.input, {
+				encoding: "UTF-8",
+			});
+			const html = render(inputText, argv.renderer);
+
+			const outputDir = path.dirname(argv.output);
+			fs.mkdirSync(outputDir, { recursive: true });
+			// Write everything to the output file
+			fs.writeFileSync(argv.output, html);
+			// Write supporting web assets
+			fs.cpSync("./build/", path.join(outputDir, "./"), { recursive: true });
+
+			console.log(`Output written to file: ${argv.output}`);
+		},
+	)
+	.help().argv;
